Consolidate AlignmentOptimizer tests into single test file
This commit simplifies the test structure by consolidating all AlignmentOptimizer
related tests into a single comprehensive test file:

- Move integration tests from test_judge_alignment_integration.py to test_alignment_optimizer.py
  - Added test_judge_align_method() and test_judge_align_method_delegation()
  - Added MockOptimizerWithTracking class for proper test isolation
  - Added create_mock_traces() helper function
  - Added necessary unittest.mock imports

- Remove tests/genai/judges/test_judge_alignment_integration.py
  - No longer needed since all tests consolidated
  - Eliminates redundant test structure
  - Simplifies test maintenance

- Maintain complete test coverage (5/5 tests still passing)
  - All abstract class behavior tests preserved
  - All integration and delegation tests preserved
  - No functionality lost in consolidation

Benefits:
- Single source for all AlignmentOptimizer tests
- Reduced test file sprawl
- Easier maintenance and discovery
- Consistent test patterns in one location
